ptracer: Use signal SIGWINCH to restrict trace
If ptracer is started with -u, then it will look for SIGWINCH signal to
detect the start and the stop of tracing.  The signals can be generated
from within the code or can be sent externally.

Most of the applications do not use SIGWINCH and the default action for
SIGWINCH is ignore, so SIGWINCH is a good candidate for re-use.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>